[Model][MiniMax-H3] Isolate temporal chunk coordination
